Refactor and simplify the energy calculations
I have never liked the energy impact calulations.
They were convoluted, copy-pasted and made my head hurt.

They also need to be split into:
- energy calculations per mode (which applies to both studies and programs), and
- energy impact, which requires the replaced mode, and only applies to programs.

@zackAmmer made some changes to split out study and program calculations but
seems to have missed this.

In this commit, we completely restructure the calculations.

The energy_intensity function simply maps the mode column (replaced mode or
mode_confirm) to the corresponding intensities

The energy_footprint multiplies it so that each trip has its footprint

The energy_impact ensures that both mode and replaced mode footprints are
computed, and computes the difference.

Not sure if this is perfect, but it is signficantly better than the morass that
we have now.

Testing done:
- generic_timeseries loads data and runs on the program environment without the replaced mode
- energy calculations before and after the scaffolding change on the CEO May dataset generate the same results